priority rendering for BG diagnols
for loop updated so that x is the driving factor, the outside left edge (value 0) is first and it draws the entire column for that X, then it goes to the max value and draws the far right column. This repeats until the middle column is drawn last. This way the center column is drawn on top of everything giving a nice brightness based overlap, since outer edges are dark and inside is the lightest.